requirements.txt
pandas==2.2.2
pdfplumber==0.11.0
camelot-py[cv]==0.11.0
redis==5.0.4
PyMuPDF
pdf2image
